refactor : secret null 이슈 리팩토링
