+++ src/main/java/com/yukgaejang/inflearnclone/domain/login/dao/AuthorityDao.java
@@ -1,4 +1,4 @@
-package com.yukgaejang.inflearnclone.domain.login.repository;
+package com.yukgaejang.inflearnclone.domain.login.dao;
 
 import com.yukgaejang.inflearnclone.domain.login.entity.Authority;
 import org.springframework.data.jpa.repository.JpaRepository;
